Configuration and main method
